openShortcutPopup
